Modernize codebase for C++23 and apply wide-ranging cleanups
- Apply C++23 modernization and related cleanups across the codebase
  (core, main, zone, term, src, tests and utilities).
- Update top-level docs and metadata: CMakeLists, README.md, SECURITY.md.
- Add CPP23 modernization documentation and helper tooling:
  docs/CPP23_MODERNIZATION*, scripts/modernize_cpp23.sh.
- Introduce utility headers and helpers (e.g. object_pool, thread_pool,
  xft_color_cache) and example cpp23 utilities under src/utils.
- Refactor and tidy term layering (`term/layer.cc` / `term/layer.hh`) and
  other modules to improve API clarity and modern C++ usage.

Summary: 193 tracked files modified, 17 new files added (approx.).

This commit groups a wide modernization sweep; follow-up commits can split
or revert specific parts if necessary.

+++ external/date/include/date/date.h
@@ -963,8 +963,8 @@ operator<<(std::basic_ostream<CharT, Traits>& os, const year_month_weekday_last&
 inline namespace literals
 {
 
-CONSTCD11 date::day  operator "" _d(unsigned long long d) NOEXCEPT;
-CONSTCD11 date::year operator "" _y(unsigned long long y) NOEXCEPT;
+CONSTCD11 date::day  operator""_d(unsigned long long d) NOEXCEPT;
+CONSTCD11 date::year operator""_y(unsigned long long y) NOEXCEPT;
 
 }  // inline namespace literals
 #endif // !defined(_MSC_VER) || (_MSC_VER >= 1900)
@@ -1972,7 +1972,7 @@ inline namespace literals
 CONSTCD11
 inline
 date::day
-operator "" _d(unsigned long long d) NOEXCEPT
+operator""_d(unsigned long long d) NOEXCEPT
 {
     return date::day{static_cast<unsigned>(d)};
 }
@@ -1980,7 +1980,7 @@ operator "" _d(unsigned long long d) NOEXCEPT
 CONSTCD11
 inline
 date::year
-operator "" _y(unsigned long long y) NOEXCEPT
+operator""_y(unsigned long long y) NOEXCEPT
 {
     return date::year(static_cast<int>(y));
 }
@@ -6502,8 +6502,8 @@ read(std::basic_istream<CharT, Traits>& is, int a0, Args&& ...args)
 {
     if (a0 != -1)
     {
-        auto u = static_cast<unsigned>(a0);
-        CharT buf[std::numeric_limits<unsigned>::digits10+2u] = {};
+        auto u = static_cast<unsigned long long>(a0);
+        CharT buf[std::numeric_limits<unsigned long long>::digits10+2u] = {};
         auto e = buf;
         do
         {
